+++ js/dashboard/Rig.js
@@ -175,7 +175,7 @@
             '<div class="stat-value">' + value + '</div>' +
             '<div class="stat-label">' + name.replace(/_|-|\./g, ' ') + '</div>' +
             '<div class="progress progress-striped">' +
-             '<div class="progress-bar progress-bar-' + progress + '" style="width: ' + share + '"></div>' +
+             '<div class="progress-bar progress-bar-' + progress + '" style="width: ' + share + '%"></div>' +
             '</div>' +
            '</div>'
   }